\name{gsa_exp}
\alias{gsa_exp}
\docType{data}
\title{
  Gene expression matrix in gene set analysis
}
\description{
  P53 NCI-60 data set provided by Subramanian \emph{et al}., (2005). A gene expression matrix whose rows correspond to genes and column correspond to samples. 
}
\usage{data(gsa_exp)}
\format{
  A matrix.
}
\source{
  http://www.broadinstitute.org/gsea
}
\references{
  Subramanian, A., Tamayo, P., Mootha, V. K., \emph{et al}. Gene set enrichment analysis: a knowledge-based approach for interpreting genome-wide expression profiles. \emph{Proceedings of the National Academy of Sciences of the United States of America}, 2005, 102(43): 15545-15550.
}
\examples{
data(gsa_exp)
}
\keyword{datasets}
